Modify default weight function
The previous default weight function was Math.random(). Now, I have exposed
`Discover.weight` as an over-ridable function for setting the default weight.

The default implementation of Discover.weight is the negative value of
`Date.now()` divided by 10 to the power of the number of digits in the
value of `Date.now`. This method will favor the earliest started nodes to be
master nodes while keeping the automatically generated weights > -1 and < 1.