Allow for checkInterval and helloInterval to be functions
It may be desirable to provide a function that can dynamically determine
what the checkInterval and helloInterval should be depending on the state
of the network. If there are lots of nodes, maybe you want the hello intervals
to be larger so there isn't as much broadcast traffic.

However, I'm just realizing in order for this to be useful, you would need to
stop() and start() in order for these changes to take place