Refactor the leadership module to make use of existing and new events
It is not totally necessary to create a whole new class/interface to
manage the leadership. With the addition of a few new events, it's
possible to wire everything up just using events. I feel like this
makes it more generic as well.

This commit removes most of the internal references to leadershipElector
and modifies the BasicLeadershipElection class to accept the node-discover
instance in the constructor. The constructor then wires all the events to
the class methods.

This restores the expected behavior that is reported in #36

ref: #36, #35, #34